refactor: 설정 페이지 토글 컴포넌트 분리
